+++ src/main/scala/edu/ie3/osmogrid/lv/region_coordinator/OsmoGridModelPartitioner.scala
@@ -42,7 +42,7 @@ object OsmoGridModelPartitioner extends LazyLogging {
     */
   def partition(
       osmoGridModel: LvOsmoGridModel,
-      areas: ParMap[AreaKey, Polygon],
+      areas: ParMap[AreaKey, Seq[Polygon]],
   ): ParMap[AreaKey, LvOsmoGridModel] = {
 
     val buildings = assign(osmoGridModel.buildings, areas, AssignByMax)
@@ -76,7 +76,7 @@ object OsmoGridModelPartitioner extends LazyLogging {
 
   private def assign(
       enhancedEntities: ParSeq[EnhancedOsmEntity],
-      areas: ParMap[AreaKey, Polygon],
+      areas: ParMap[AreaKey, Seq[Polygon]],
       allocationStrategy: EntityAllocationStrategy,
   ): ParMap[AreaKey, ParSeq[EnhancedOsmEntity]] = {
     enhancedEntities
@@ -91,7 +91,7 @@ object OsmoGridModelPartitioner extends LazyLogging {
 
   private def assign(
       enhancedEntity: EnhancedOsmEntity,
-      areas: ParMap[AreaKey, Polygon],
+      areas: ParMap[AreaKey, Seq[Polygon]],
       allocationStrategy: EntityAllocationStrategy,
   ): Iterable[AreaKey] = {
     val entityVotes = vote(enhancedEntity.entity, enhancedEntity, areas)
@@ -117,7 +117,7 @@ object OsmoGridModelPartitioner extends LazyLogging {
   private def vote(
       entity: OsmEntity,
       enhancedEntity: EnhancedOsmEntity,
-      areas: ParMap[AreaKey, Polygon],
+      areas: ParMap[AreaKey, Seq[Polygon]],
   ): Map[AreaKey, Int] =
     entity match {
       case node: OsmEntity.Node =>
@@ -131,7 +131,7 @@ object OsmoGridModelPartitioner extends LazyLogging {
   private def vote(
       relation: OsmEntity.Relation,
       enhancedEntity: EnhancedOsmEntity,
-      areas: ParMap[AreaKey, Polygon],
+      areas: ParMap[AreaKey, Seq[Polygon]],
   ): Map[AreaKey, Int] =
     relation.members
       .flatMap { member =>
@@ -157,7 +157,7 @@ object OsmoGridModelPartitioner extends LazyLogging {
   private def vote(
       way: OsmEntity.Way,
       enhancedEntity: EnhancedOsmEntity,
-      areas: ParMap[AreaKey, Polygon],
+      areas: ParMap[AreaKey, Seq[Polygon]],
   ): Map[AreaKey, Int] =
     way.nodes
       .flatMap(enhancedEntity.node)
@@ -171,14 +171,12 @@ object OsmoGridModelPartitioner extends LazyLogging {
 
   private def vote(
       node: OsmEntity.Node,
-      areas: ParMap[AreaKey, Polygon],
+      areas: ParMap[AreaKey, Seq[Polygon]],
   ): Map[AreaKey, Int] = {
     val point = GeoUtils.buildPoint(node.latitude, node.longitude)
-    areas.iterator
-      .collectFirst {
-        case areaId -> polygon if polygon.covers(point) =>
-          Map(areaId -> 1)
-      }
-      .getOrElse(Map.empty)
+    areas.iterator.collect {
+      case (areaId, polygons) if polygons.exists(_.covers(point)) =>
+        areaId -> 1
+    }.toMap
   }
 }