+++ src/test/scala/edu/ie3/osmogrid/lv/region_coordinator/BoundaryFactorySpec.scala
@@ -9,16 +9,37 @@ package edu.ie3.osmogrid.lv.region_coordinator
 import edu.ie3.osmogrid.io.input.BoundaryAdminLevel
 import edu.ie3.test.common.UnitSpec
 
+import scala.collection.parallel.ParSet
 import scala.io.Source
+import scala.math.BigDecimal.RoundingMode
 import scala.util.Using
 
 class BoundaryFactorySpec extends UnitSpec {
 
   private val osmoGridModel = LvTestModel.osmoGridModel
+  private val osmoGridModelHelgoland = LvTestModel.osmoGridModelHelgoland
 
-  // polygons can be manually validated e.g. in QGIS
   "Creating boundary polygons from osm data" when {
     "having proper input data" should {
+      "result in correct polygons on municipality level with multiple polygons" in {
+        val rawPolygons = getFileLines("Helgoland_boundaries_level8.csv")
+        val polygons = rawPolygons.map(truncate(_, 6))
+
+        val actualPolygons: ParSet[String] =
+          BoundaryFactory
+            .buildBoundaryPolygons(
+              osmoGridModelHelgoland,
+              BoundaryAdminLevel.MUNICIPALITY_LEVEL,
+            )
+            .values
+            .flatMap(_.toList)
+            .map(_.toString)
+            .map(truncate(_, 6))
+            .toSet
+
+        actualPolygons.mkString(", ") shouldBe polygons.mkString(", ")
+      }
+
       "result in correct polygons on county level" in {
         val polygons = getFileLines("DoBoCas_boundaries_level6.csv")
 
@@ -29,11 +50,15 @@ class BoundaryFactorySpec extends UnitSpec {
               BoundaryAdminLevel.COUNTY_LEVEL,
             )
             .values
+            .flatMap(_.toList) // Flatten the lists of polygons
             .map(_.toString)
-            .seq
             .toSet
 
-        actualPolygons shouldBe polygons
+        val expectedSet = polygons.map(_.trim)
+        val actualSet = actualPolygons.map(_.trim)
+
+        actualSet.size shouldBe expectedSet.size
+        actualSet.forall(expectedSet.contains) shouldBe true
       }
 
       "result in correct polygons on municipality level" in {
@@ -70,6 +95,20 @@ class BoundaryFactorySpec extends UnitSpec {
         actualPolygons shouldBe polygons
       }
     }
+  }
+
+  private def truncate(polygon: String, precision: Int): String = {
+    val pattern = "(-?\\d+\\.\\d+) (-?\\d+\\.\\d+)".r
+    pattern.replaceAllIn(
+      polygon,
+      m =>
+        BigDecimal(m.group(1))
+          .setScale(precision, RoundingMode.DOWN)
+          .toString() + " " +
+          BigDecimal(m.group(2))
+            .setScale(precision, RoundingMode.DOWN)
+            .toString(),
+    )
   }
 
   private def getFileLines(resource: String): Set[String] = {